implemented cleanup
